+++ Assets/HUF/AnalyticsAppsFlyer/SDK/AppsFlyer/Plugins/iOS/AppsFlyeriOSWrapper.mm
@@ -96,7 +96,7 @@ extern "C" {
 
     const void _setOneLinkCustomDomains (int length, const char **oneLinkCustomDomains) {
         if(length > 0 && oneLinkCustomDomains) {
-            [[AppsFlyerLib shared] setResolveDeepLinkURLs:NSArrayFromCArray(length, oneLinkCustomDomains)];
+            [[AppsFlyerLib shared] setOneLinkCustomDomains:NSArrayFromCArray(length, oneLinkCustomDomains)];
         }
     }